Add controller endpoint for sentry project list
Also modify the project_error_tracking_settings model to validate
that api_url and token are present when enabled is true.
Also set the default of enabled column to false and remove enabled from
the list of columns to be exported so that when the data is imported,
enabled will default to false. This will prevent the validation from
being triggered since the token is not exported and will not be present
after import.