Add migration to add columns to error_tracking
Add project_name and organization_name to hold the names of the selected
project and its organization.
Change the default of the enabled column to be false, so that on project
import, the enabled column will default to false and will not cause the
validations to fail since the token will not be present.
Also allow api_url to be nil, since it can be nil if enabled is false.