Make GraphQL Types authorizable
Added a new initializer to allow the `authorize` keyword to be used
when defining a field, or a Type.

Modified Gitlab::GraphQL::Authorize and associated classes to check the
authorize data from the authorize metadata, and support checks on Types,
as well as Scalar fields.

Moved existing checks for Types from their field definitions and into
the Type definition.

Issue:
https://gitlab.com/gitlab-org/gitlab-ce/issues/54417