Fix failing e2e labels test
On staging new projects now have existing labels, which displays a
different layout causing a test to fail because an expexted image
is not displayed.

There is a similar image shown on the page with existing labels, so
this change adds a selector so the test can successfully wait for the
image to appear.