Handle error for #batch_add_keys
Since there are 2 different branches for #branch_add_keys,
we need to handle 2 different errors but ensure that it
returns the same value (true or false).

This way, the consumer of the method can handle whatever
the case is easily. In this case, it's only being used
in `gitlab:shell:setup`.