Move permissions check to Entity::Release
Instead of relying on endpoint to generate the response based on the
current user's access level pass the user to the release API entity and
do the check there.

Use download_code instead of read_commit_status in order to determine
can the current user see tag and commit information.